May 2023 - Major update
New components added:
- CAVA
- Hyprland
- Hyprpaper
- Waybar

Extra components updated:
- Portage
- ZSH
- VS Code
- Picom
- i3

Also, deleted Gamemode and `.zprofile`. From now on it is better to
access the Pingu DE through the X.Org server (with `startx` command), or
through the Hyprland Wayland compositor (with `startw` command). It
should be noted that the Hyprland configuration and setup is still being
highly developed, so the `X.Org + i3 + picom` is the one to be used for
production.