fix: `hid-playstation` driver for DS4 instead of `hid-sony`
With the Linux 6.2 kernel Sony DualShock 4 controller support was added to the "hid-playstation" driver as the newer open-source driver maintained by Sony and started out originally in supporting the PlayStation 5 DualSense controllers. The DualShock 4 controller was long supported by the "hid-sony" driver while now in the Linux 6.3 kernel that support will be removed.

Sony introduced the hid-playstation Linux driver two years ago for their newer PlayStation controllers and being modern and up to their software/driver standards over hid-sony that has long been maintained by the community.

Now that the DualShock 4 controller support is added to the hid-playstation driver, they are removing that controller support from hid-sony for Linux 6.3. Many were wondering if this was going to happen since normally one piece of hardware isn't supported by two different drivers within the Linux kernel.

With this merge to the HID subsystem's "for-next" branch, that DualShock 4 controller support is cleared out from hid-sony when it comes to the Linux 6.3 merge window.

**More info:** <https://www.phoronix.com/news/HID-Sony-Dropping-DS4>